# Optimize listAll query