# Summary

Improve list of models page performance
Benchmark result before and after executing the same request:

| Package | Before | After |
| -------- | ------- | ----- |
| size | 421kb | 27kb | 
| Response time | 180ms | 70ms |